fixed sound effect and PlayerResult in server

+++ Trivia/Trivia/PlayerResults.h
@@ -5,5 +5,5 @@ struct PlayerResults {
 	std::string username;
 	unsigned int correctAnswerCount;
 	unsigned int wrongAnswerCount;
-	unsigned int averageAnswerTime;
+	float averageAnswerTime;
 };
